Issue-52: Created core package and move there all common classes for user UI. Fixed formatting.
