+++ app/src/main/java/com/kirchhoff/movies/network/services/MovieService.kt
@@ -3,7 +3,6 @@ package com.kirchhoff.movies.network.services
 import com.kirchhoff.movies.data.responses.DiscoverMoviesResponse
 import com.kirchhoff.movies.data.responses.MovieDetails
 import com.kirchhoff.movies.data.responses.ReviewsListResponse
-import kotlinx.coroutines.Deferred
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
