+++ app/src/main/java/com/kirchhoff/movies/ui/screens/core/PaginatedScreenVM.kt
@@ -1,4 +1,4 @@
-package com.kirchhoff.movies.ui.screens
+package com.kirchhoff.movies.ui.screens.core
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
