+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/PersonsVM.kt
@@ -2,7 +2,7 @@ package com.kirchhoff.movies.ui.screens.main.persons
 
 import com.kirchhoff.movies.data.responses.PersonsResponse
 import com.kirchhoff.movies.repository.person.IPersonsRepository
-import com.kirchhoff.movies.ui.screens.PaginatedScreenVM
+import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
 
 class PersonsVM(private val personRepository: IPersonsRepository) :
     PaginatedScreenVM<PersonsResponse>() {
