+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/tvs/TvsFragment.kt
@@ -2,8 +2,8 @@ package com.kirchhoff.movies.ui.screens.main.tvs
 
 import com.kirchhoff.movies.data.Tv
 import com.kirchhoff.movies.data.responses.DiscoverTvsResponse
+import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.details.DetailsActivity
-import com.kirchhoff.movies.ui.screens.PaginatedScreenFragment
 import com.kirchhoff.movies.ui.screens.main.tvs.adapter.TvsListAdapter
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
 import org.koin.android.viewmodel.ext.android.viewModel
