+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/tvs/TvsVM.kt
@@ -2,7 +2,7 @@ package com.kirchhoff.movies.ui.screens.main.tvs
 
 import com.kirchhoff.movies.data.responses.DiscoverTvsResponse
 import com.kirchhoff.movies.repository.discover.IDiscoverRepository
-import com.kirchhoff.movies.ui.screens.PaginatedScreenVM
+import com.kirchhoff.movies.ui.screens.core.PaginatedScreenVM
 
 class TvsVM(private val discoverRepository: IDiscoverRepository) :
     PaginatedScreenVM<DiscoverTvsResponse>() {
