+++ app/src/main/java/com/kirchhoff/movies/ui/screens/similar/movie/SimilarMoviesFragment.kt
@@ -3,14 +3,13 @@ package com.kirchhoff.movies.ui.screens.similar.movie
 import android.os.Bundle
 import com.kirchhoff.movies.data.Movie
 import com.kirchhoff.movies.data.responses.DiscoverMoviesResponse
-import com.kirchhoff.movies.ui.screens.PaginatedScreenFragment
+import com.kirchhoff.movies.ui.screens.core.PaginatedScreenFragment
+import com.kirchhoff.movies.ui.screens.core.movies.adapter.MoviesListAdapter
 import com.kirchhoff.movies.ui.screens.details.DetailsActivity
-import com.kirchhoff.movies.ui.screens.details.movie.MovieDetailsFragment
-import com.kirchhoff.movies.ui.screens.main.movies.adapter.MoviesListAdapter
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseRecyclerViewAdapter
 import org.koin.android.viewmodel.ext.android.viewModel
 
-class SimilarMoviesFragment  : PaginatedScreenFragment<Movie, DiscoverMoviesResponse>(),
+class SimilarMoviesFragment : PaginatedScreenFragment<Movie, DiscoverMoviesResponse>(),
     BaseRecyclerViewAdapter.OnItemClickListener<Movie> {
 
     override val vm by viewModel<SimilarMoviesVM>()
